requirements.txt
Django
requests
mysqlclient
djangorestframework
pygments
gensim
nltk
django-cors-headers
pandas
matplotlib
scikit-learn